requirements.txt
streamlit==1.31.0
PyPDF2==3.0.1
python-docx==0.8.11
